update / 2.0 incoming (only !ping and memes work for now)
