+++ modules/commands/jff/ahoj.js
@@ -1,0 +1,8 @@
+
+const COLORS = require("../../consts").COLORS;
+
+module.exports = {
+    command: function(msg) {
+        msg.reply("Ahoj");  //robil Dan Valnicek
+    }
+}
