+++ modules/commands/jff/bye.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+    command: function(msg) {
+        msg.channel.send("I'll be waiting...");
+    }
+}
