+++ modules/commands/jff/rip.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+    command: function(msg) {
+        msg.channel.send("Rest in piss, forever miss...");
+    }
+}
