+++ modules/commands/jff/skap.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+    command: function(msg) {
+        msg.channel.send("``` ((skap)^2) / hned ```");
+    }
+}
