+++ modules/commands/jff/zhni.js
@@ -1,0 +1,8 @@
+
+const COLORS = require("../../consts").COLORS;
+
+module.exports = {
+    command: function(msg) {
+        msg.channel.send("Sa najedz ked si zhni lol");
+    }
+}
