+++ modules/db.js
@@ -58,7 +58,7 @@ module.exports = {
 
     save: (data)=>{
         if (process.env.DISABLE_SAVE == "yes") {return;} // for beta
-
+        return;
         console.log("[SAVE] Saving events...");
 
         if ((Object.keys(usersObj).length === 0 && usersObj.constructor === Object) || (events.length < 1) || (!data)) {
